requirements.txt
requests>=2.31.0
flask>=2.3.0
python-dotenv>=1.0.0
aiohttp>=3.8.0
xmltodict>=0.13.0
flask-swagger-ui>=4.11.1
gunicorn>=21.0.0
flask-cors>=4.0.0
flask-limiter>=3.5.0
gevent>=23.0.0
prometheus-client>=0.17.0
psutil>=5.9.0 